+++ fetching-service/src/main/java/rs/edu/raf/banka1/services/FuturesService.java
@@ -14,4 +14,6 @@ public interface FuturesService {
     int addAllFutures(List<ListingFuture> futures);
 
     Boolean addFuture(ListingFuture future);
+
+    Optional<ListingFuture> findByTicker(String ticker);
 }
