+++ fetching-service/src/test/java/rs/edu/raf/banka1/services/ListingStockServiceImplTest.java
@@ -1,48 +1,87 @@
 package rs.edu.raf.banka1.services;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ArrayNode;
+import org.junit.jupiter.api.Disabled;
+import  org.springframework.core.io.Resource;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
+
+import org.mockito.*;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.core.io.Resource;
 import rs.edu.raf.banka1.mapper.StockMapper;
 import rs.edu.raf.banka1.model.ListingHistory;
 import rs.edu.raf.banka1.model.ListingStock;
-import rs.edu.raf.banka1.repositories.ListingHistoryRepository;
-import rs.edu.raf.banka1.repositories.StockRepository;
+import rs.edu.raf.banka1.model.entities.Country;
+import rs.edu.raf.banka1.model.entities.Exchange;
+import rs.edu.raf.banka1.model.entities.Holiday;
+import rs.edu.raf.banka1.repositories.*;
+import rs.edu.raf.banka1.utils.Requests;
+
 
 import java.sql.Date;
-import java.time.ZoneOffset;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.*;
+import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.when;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 public class ListingStockServiceImplTest {
-    @Mock
+
     private ListingHistoryRepository listingHistoryRepository;
-    @Mock
-    private StockMapper listingMapper;
-    @Mock
+
     private StockRepository stockRepository;
-    @Mock
+
+    private ExchangeRepository exchangeRepository;
+
     private StockMapper stockMapper;
-    @InjectMocks
+
+    private ObjectMapper objectMapper;
+
+    @Mock
+    private Resource resource;
+
+    @Mock
+    private JsonNode jsonNode;
+
+
     private ListingStockServiceImpl listingStockService;
+
     private ListingStock stockAAPL;
     private ListingStock stockMSFT;
+    private ListingStock stockDT;
     private List<ListingStock> stocks;
+    private Exchange exchangeDT;
+    private Country countryUS;
 
     private List<ListingHistory> lst = new ArrayList<>();
     private ListingHistory model1;
     private ListingHistory model2;
     private long date;
+
     @BeforeEach
     public void setUp(){
+
+        listingHistoryRepository = mock(ListingHistoryRepository.class);
+        stockRepository = mock(StockRepository.class);
+        exchangeRepository = mock(ExchangeRepository.class);
+        stockMapper = mock(StockMapper.class);
+        listingStockService = new ListingStockServiceImpl();
+        listingStockService.setStockRepository(stockRepository);
+        listingStockService.setListingHistoryRepository(listingHistoryRepository);
+        listingStockService.setExchangeRepository(exchangeRepository);
+        listingStockService.setStockMapper(stockMapper);
+
+        objectMapper = new ObjectMapper();
+
         // stock data
         stockAAPL = new ListingStock();
         stockAAPL.setTicker("AAPL");
@@ -52,9 +91,45 @@ public class ListingStockServiceImplTest {
         stockMSFT.setTicker("MSFT");
         stockMSFT.setPrice(200.0);
 
+        SimpleDateFormat hoursDateFormat = new SimpleDateFormat("HH:mm:ss");
+        countryUS = new Country();
+        countryUS.setISOCode("US");
+        countryUS.setId(113L);
+        countryUS.setTimezoneOffset(0);
+        try {
+            countryUS.setOpenTime(new java.util.Date(hoursDateFormat.parse("08:00:00").getTime()));
+            countryUS.setCloseTime(new java.util.Date(hoursDateFormat.parse("18:00:00").getTime()));
+
+        } catch (ParseException e) {
+            throw new RuntimeException(e);
+        }
+
+        exchangeDT = new Exchange();
+        exchangeDT.setId(2295L);
+        exchangeDT.setExchangeName("NEW YORK STOCK EXCHANGE, INC.");
+        exchangeDT.setMicCode("XNYS");
+        exchangeDT.setCountry(countryUS);
+
+        stockDT = new ListingStock();
+        stockDT.setTicker("DT");
+        stockDT.setName("Dynatrace Holdings LLC");
+        stockDT.setListingType("Stock");
+        stockDT.setExchangeName("NEW YORK STOCK EXCHANGE, INC.");
+        stockDT.setListingId(1L);
+        stockDT.setExchange(exchangeDT);
+        stockDT.setVolume(2566374);
+        stockDT.setPriceChange(-0.705);
+        stockDT.setPrice(45.735);
+        stockDT.setOutstandingShares(295999000);
+        stockDT.setLow(45.31);
+        stockDT.setLastRefresh(1712005154);
+        stockDT.setHigh(46.29);
+        stockDT.setDividendYield((double) 0);
+
         stocks = new ArrayList<>();
         stocks.add(stockAAPL);
         stocks.add(stockMSFT);
+        stocks.add(stockDT);
 
         // history data
         model1 = new ListingHistory();
@@ -121,4 +196,377 @@ public class ListingStockServiceImplTest {
         assertEquals(lst.size(), listingStockService.addAllListingsToHistory(lst));
     }
 
+
+    @Test
+    public void testFetchSingleListingHistory() {
+        // Arrange
+        String ticker = "AAPL";
+        String apiUrl = listingStockService.getHistoryListingApiUrl() + ticker + "&outputsize=compact&apikey=" + listingStockService.getListingAPItoken();
+        String response = "{\n" +
+                "    \"Meta Data\": {\n" +
+                "        \"1. Information\": \"Daily Prices (open, high, low, close) and Volumes\",\n" +
+                "        \"2. Symbol\": \"AAPL\",\n" +
+                "        \"3. Last Refreshed\": \"2024-04-12\",\n" +
+                "        \"4. Output Size\": \"Compact\",\n" +
+                "        \"5. Time Zone\": \"US/Eastern\"\n" +
+                "    },\n" +
+                "    \"Time Series (Daily)\": {\n" +
+                "        \"2024-04-12\": {\n" +
+                "            \"1. open\": \"174.2600\",\n" +
+                "            \"2. high\": \"178.3600\",\n" +
+                "            \"3. low\": \"174.2100\",\n" +
+                "            \"4. close\": \"176.5500\",\n" +
+                "            \"5. volume\": \"101670886\"\n" +
+                "        },\n" +
+                "        \"2024-04-11\": {\n" +
+                "            \"1. open\": \"168.3400\",\n" +
+                "            \"2. high\": \"175.4600\",\n" +
+                "            \"3. low\": \"168.1600\",\n" +
+                "            \"4. close\": \"175.0400\",\n" +
+                "            \"5. volume\": \"91070275\"\n" +
+                "        }}}";
+
+        try(MockedStatic<Requests> req = Mockito.mockStatic(Requests.class)) {
+            req.when(() -> Requests.sendRequest(any())).thenReturn(response);
+            List<ListingHistory> result = listingStockService.fetchSingleListingHistory(ticker);
+
+            assertEquals(2, result.size());
+        }
+
+    }
+
+    @Test
+    public void addListingStockNotPresentTest(){
+        when(stockRepository.findByTicker("AAPL")).thenReturn(Optional.empty());
+        assertEquals(1, listingStockService.addListingStock(stockAAPL));
+    }
+
+    @Test
+    public void addListingStockPresentTest(){
+        ListingStock updateStock = new ListingStock();
+        updateStock.setTicker("AAPL");
+        updateStock.setPrice(101.0);
+        when(listingStockService.findByTicker("AAPL")).thenReturn(Optional.of(stockAAPL));
+        assertEquals(0, listingStockService.addListingStock(updateStock));
+    }
+
+    @Test
+    public void addAllListingStocksPresentTests(){
+        when(listingStockService.findByTicker("AAPL")).thenReturn(Optional.of(stockAAPL));
+        when(listingStockService.findByTicker("MSFT")).thenReturn(Optional.of(stockMSFT));
+        assertEquals(1, listingStockService.addAllListingStocks(stocks));
+    }
+
+    // puca test
+    @Test
+    @Disabled
+    public void fetchNListingStocks() {
+//        Resource resource = new ClassPathResource(Constants.listingsFilePath,this.getClass().getClassLoader());
+
+        String baseResponse = "{\n" +
+                "    \"Global Quote\": {\n" +
+                "        \"01. symbol\": \"CDLX\",\n" +
+                "        \"02. open\": \"14.1800\",\n" +
+                "        \"03. high\": \"14.2300\",\n" +
+                "        \"04. low\": \"13.5500\",\n" +
+                "        \"05. price\": \"13.5700\",\n" +
+                "        \"06. volume\": \"650480\",\n" +
+                "        \"07. latest trading day\": \"2024-04-12\",\n" +
+                "        \"08. previous close\": \"14.3800\",\n" +
+                "        \"09. change\": \"-0.8100\",\n" +
+                "        \"10. change percent\": \"-5.6328%\"\n" +
+                "    }\n" +
+                "}";
+        String updateResponse = "{\n" +
+                "    \"Symbol\": \"CDLX\",\n" +
+                "    \"AssetType\": \"Common Stock\",\n" +
+                "    \"Name\": \"Cardlytics Inc\",\n" +
+                "    \"Description\": \"Cardlytics, Inc. operates an advertising platform within financial institutions' digital channels including online, mobile, email, and various real-time notifications in the United States and the United Kingdom. The company is headquartered in Atlanta, Georgia.\",\n" +
+                "    \"CIK\": \"1666071\",\n" +
+                "    \"Exchange\": \"NASDAQ\",\n" +
+                "    \"Currency\": \"USD\",\n" +
+                "    \"Country\": \"USA\",\n" +
+                "    \"Sector\": \"TECHNOLOGY\",\n" +
+                "    \"Industry\": \"SERVICES-COMPUTER PROGRAMMING, DATA PROCESSING, ETC.\",\n" +
+                "    \"Address\": \"675 PONCE DE LEON AVENUE, NE, SUITE 6000, ATLANTA, GA, US\",\n" +
+                "    \"FiscalYearEnd\": \"December\",\n" +
+                "    \"LatestQuarter\": \"2023-12-31\",\n" +
+                "    \"MarketCapitalization\": \"598560000\",\n" +
+                "    \"EBITDA\": \"-37209000\",\n" +
+                "    \"PERatio\": \"None\",\n" +
+                "    \"PEGRatio\": \"None\",\n" +
+                "    \"BookValue\": \"3.393\",\n" +
+                "    \"DividendPerShare\": \"None\",\n" +
+                "    \"DividendYield\": \"None\",\n" +
+                "    \"EPS\": \"-3.69\",\n" +
+                "    \"RevenuePerShareTTM\": \"8.47\",\n" +
+                "    \"ProfitMargin\": \"-0.436\",\n" +
+                "    \"OperatingMarginTTM\": \"-0.0867\",\n" +
+                "    \"ReturnOnAssetsTTM\": \"-0.0629\",\n" +
+                "    \"ReturnOnEquityTTM\": \"-0.778\",\n" +
+                "    \"RevenueTTM\": \"309204000\",\n" +
+                "    \"GrossProfitTTM\": \"115415000\",\n" +
+                "    \"DilutedEPSTTM\": \"-3.69\",\n" +
+                "    \"QuarterlyEarningsGrowthYOY\": \"-0.567\",\n" +
+                "    \"QuarterlyRevenueGrowthYOY\": \"0.081\",\n" +
+                "    \"AnalystTargetPrice\": \"17.67\",\n" +
+                "    \"AnalystRatingStrongBuy\": \"0\",\n" +
+                "    \"AnalystRatingBuy\": \"2\",\n" +
+                "    \"AnalystRatingHold\": \"1\",\n" +
+                "    \"AnalystRatingSell\": \"0\",\n" +
+                "    \"AnalystRatingStrongSell\": \"0\",\n" +
+                "    \"TrailingPE\": \"-\",\n" +
+                "    \"ForwardPE\": \"-\",\n" +
+                "    \"PriceToSalesRatioTTM\": \"1.936\",\n" +
+                "    \"PriceToBookRatio\": \"4.44\",\n" +
+                "    \"EVToRevenue\": \"2.499\",\n" +
+                "    \"EVToEBITDA\": \"-7.81\",\n" +
+                "    \"Beta\": \"1.466\",\n" +
+                "    \"52WeekHigh\": \"20.52\",\n" +
+                "    \"52WeekLow\": \"4.94\",\n" +
+                "    \"50DayMovingAverage\": \"10.15\",\n" +
+                "    \"200DayMovingAverage\": \"10.65\",\n" +
+                "    \"SharesOutstanding\": \"44109100\",\n" +
+                "    \"DividendDate\": \"None\",\n" +
+                "    \"ExDividendDate\": \"None\"\n" +
+                "}";
+        String baseUrl = "https://www.alphavantage.co/query?function=GLOBAL_QUOTE&symbol=CDLX&apikey=OF6BVKZOCXWHD9NS";
+        String updateUrl = "https://www.alphavantage.co/query?function=OVERVIEW&symbol=CDLX&apikey=OF6BVKZOCXWHD9NS";
+
+        ObjectMapper objectMapper1 = new ObjectMapper();
+        listingStockService.setObjectMapper(objectMapper1);
+
+        StockMapper stockMapper1 = new StockMapper();
+        listingStockService.setStockMapper(stockMapper1);
+        int n = 3;
+
+
+        try (MockedStatic<Requests> req = Mockito.mockStatic(Requests.class)) {
+            req.when(() -> Requests.sendRequest(startsWith("https://www.alphavantage.co/query?function=GLOBAL_QUOTE"))).thenReturn(baseResponse);
+            req.when(() -> Requests.sendRequest(startsWith("https://www.alphavantage.co/query?function=OVERVIEW"))).thenReturn(updateResponse);
+
+            when(exchangeRepository.findByExchangeName(any())).thenReturn(exchangeDT);
+            when(exchangeRepository.findByExchangeName(any())).thenReturn(exchangeDT);
+            when(exchangeRepository.findByExchangeName(any())).thenReturn(exchangeDT);
+
+            List<ListingStock> stocks = listingStockService.fetchNListingStocks(n);
+            assertEquals(3, stocks.size());
+        }
+    }
+
+    @Test
+    public void testFetchNStocks() {
+        List<ListingStock> allStocks = new ArrayList<>();
+        allStocks.add(new ListingStock());
+        allStocks.add(new ListingStock());
+
+        when(listingStockService.getAllStocks()).thenReturn(allStocks);
+
+        int n = 2;
+        List<ListingStock> fetchedStocks = listingStockService.fetchNStocks(n);
+
+        assertEquals(n, fetchedStocks.size());
+    }
+
+    // puca test
+    @Test
+    @Disabled
+    public void testCreateListingStock() {
+        // Arrange
+        String baseResponse = "{\n" +
+                "    \"Global Quote\": {\n" +
+                "        \"01. symbol\": \"CDLX\",\n" +
+                "        \"02. open\": \"14.1800\",\n" +
+                "        \"03. high\": \"14.2300\",\n" +
+                "        \"04. low\": \"13.5500\",\n" +
+                "        \"05. price\": \"13.5700\",\n" +
+                "        \"06. volume\": \"650480\",\n" +
+                "        \"07. latest trading day\": \"2024-04-12\",\n" +
+                "        \"08. previous close\": \"14.3800\",\n" +
+                "        \"09. change\": \"-0.8100\",\n" +
+                "        \"10. change percent\": \"-5.6328%\"\n" +
+                "    }\n" +
+                "}";
+        String updateResponse = "{\n" +
+                "    \"Symbol\": \"CDLX\",\n" +
+                "    \"AssetType\": \"Common Stock\",\n" +
+                "    \"Name\": \"Cardlytics Inc\",\n" +
+                "    \"Description\": \"Cardlytics, Inc. operates an advertising platform within financial institutions' digital channels including online, mobile, email, and various real-time notifications in the United States and the United Kingdom. The company is headquartered in Atlanta, Georgia.\",\n" +
+                "    \"CIK\": \"1666071\",\n" +
+                "    \"Exchange\": \"NASDAQ\",\n" +
+                "    \"Currency\": \"USD\",\n" +
+                "    \"Country\": \"USA\",\n" +
+                "    \"Sector\": \"TECHNOLOGY\",\n" +
+                "    \"Industry\": \"SERVICES-COMPUTER PROGRAMMING, DATA PROCESSING, ETC.\",\n" +
+                "    \"Address\": \"675 PONCE DE LEON AVENUE, NE, SUITE 6000, ATLANTA, GA, US\",\n" +
+                "    \"FiscalYearEnd\": \"December\",\n" +
+                "    \"LatestQuarter\": \"2023-12-31\",\n" +
+                "    \"MarketCapitalization\": \"598560000\",\n" +
+                "    \"EBITDA\": \"-37209000\",\n" +
+                "    \"PERatio\": \"None\",\n" +
+                "    \"PEGRatio\": \"None\",\n" +
+                "    \"BookValue\": \"3.393\",\n" +
+                "    \"DividendPerShare\": \"None\",\n" +
+                "    \"DividendYield\": \"None\",\n" +
+                "    \"EPS\": \"-3.69\",\n" +
+                "    \"RevenuePerShareTTM\": \"8.47\",\n" +
+                "    \"ProfitMargin\": \"-0.436\",\n" +
+                "    \"OperatingMarginTTM\": \"-0.0867\",\n" +
+                "    \"ReturnOnAssetsTTM\": \"-0.0629\",\n" +
+                "    \"ReturnOnEquityTTM\": \"-0.778\",\n" +
+                "    \"RevenueTTM\": \"309204000\",\n" +
+                "    \"GrossProfitTTM\": \"115415000\",\n" +
+                "    \"DilutedEPSTTM\": \"-3.69\",\n" +
+                "    \"QuarterlyEarningsGrowthYOY\": \"-0.567\",\n" +
+                "    \"QuarterlyRevenueGrowthYOY\": \"0.081\",\n" +
+                "    \"AnalystTargetPrice\": \"17.67\",\n" +
+                "    \"AnalystRatingStrongBuy\": \"0\",\n" +
+                "    \"AnalystRatingBuy\": \"2\",\n" +
+                "    \"AnalystRatingHold\": \"1\",\n" +
+                "    \"AnalystRatingSell\": \"0\",\n" +
+                "    \"AnalystRatingStrongSell\": \"0\",\n" +
+                "    \"TrailingPE\": \"-\",\n" +
+                "    \"ForwardPE\": \"-\",\n" +
+                "    \"PriceToSalesRatioTTM\": \"1.936\",\n" +
+                "    \"PriceToBookRatio\": \"4.44\",\n" +
+                "    \"EVToRevenue\": \"2.499\",\n" +
+                "    \"EVToEBITDA\": \"-7.81\",\n" +
+                "    \"Beta\": \"1.466\",\n" +
+                "    \"52WeekHigh\": \"20.52\",\n" +
+                "    \"52WeekLow\": \"4.94\",\n" +
+                "    \"50DayMovingAverage\": \"10.15\",\n" +
+                "    \"200DayMovingAverage\": \"10.65\",\n" +
+                "    \"SharesOutstanding\": \"44109100\",\n" +
+                "    \"DividendDate\": \"None\",\n" +
+                "    \"ExDividendDate\": \"None\"\n" +
+                "}";
+        String baseUrl = "https://www.alphavantage.co/query?function=GLOBAL_QUOTE&symbol=CDLX&apikey=OF6BVKZOCXWHD9NS";
+        String updateUrl = "https://www.alphavantage.co/query?function=OVERVIEW&symbol=CDLX&apikey=OF6BVKZOCXWHD9NS";
+        ListingStock stock = new ListingStock();
+        String symbol = "CDLX";
+        String companyName = "Cardlytics Inc";
+        String primaryExchange = "EURONEXT  UK - REPORTING SERVICES";
+        stock.setTicker(symbol);
+        stock.setName(companyName);
+        stock.setExchangeName(primaryExchange);
+        stock.setListingId(90L);
+        Double high = 200.0;
+        stock.setHigh(high);
+        stock.setLow(100.0);
+        stock.setPrice(150.0);
+        stock.setPriceChange(10.0);
+        stock.setVolume(100);
+        stock.setDividendYield(90.0);
+        stock.setOutstandingShares(90);
+
+
+
+        try (MockedStatic<Requests> req = Mockito.mockStatic(Requests.class)) {
+            req.when(() -> Requests.sendRequest(eq(baseUrl))).thenReturn(baseResponse);
+            req.when(() -> Requests.sendRequest(eq(updateUrl))).thenReturn(updateResponse);
+
+            when(exchangeRepository.findByExchangeName(anyString())).thenReturn(exchangeDT);
+            when(stockMapper.createListingStock(anyString(),anyString(),any(),anyDouble(),anyDouble(),anyDouble(),anyDouble(),anyInt(),anyInt(),anyDouble())).thenReturn(stock);
+            ListingStock result = listingStockService.createListingStock(symbol, companyName, primaryExchange);
+
+            // Assert
+            verify(stockMapper,times(1)).createListingStock( anyString(),anyString(),any(),anyDouble(),anyDouble(),anyDouble(),anyDouble(),anyInt(),anyInt(),anyDouble());
+            assertEquals(high,200);
+            assertEquals(symbol, result.getTicker());
+            assertEquals(companyName, result.getName());
+            assertEquals(primaryExchange, result.getExchangeName());
+
+        }
+
+    }
+
+    @Test
+    public void testReformatNamesToJSON() throws Exception {
+        // Arrange
+        String response = "[{\"symbol\":\"AAPL\",\"companyName\":\"Apple Inc.\",\"primaryExchange\":\"NASDAQ\"}," +
+                "{\"symbol\":\"MSFT\",\"companyName\":\"Microsoft Corporation\",\"primaryExchange\":\"NASDAQ\"}]";
+
+        // Act
+        ArrayNode result = listingStockService.reformatNamesToJSON(response);
+
+        // Assert
+        assertEquals(2, result.size());
+
+        JsonNode firstNode = result.get(0);
+        assertEquals("AAPL", firstNode.get("symbol").asText());
+        assertEquals("Apple Inc.", firstNode.get("companyName").asText());
+        assertEquals("NASDAQ", firstNode.get("primaryExchange").asText());
+
+        JsonNode secondNode = result.get(1);
+        assertEquals("MSFT", secondNode.get("symbol").asText());
+        assertEquals("Microsoft Corporation", secondNode.get("companyName").asText());
+        assertEquals("NASDAQ", secondNode.get("primaryExchange").asText());
+    }
+
+    @Test
+    public void fetchNListingsHistoryTest(){
+        ListingStock stock1 = new ListingStock();
+        stock1.setListingId(9L);
+        stock1.setTicker("IBM");
+        ListingStock stock2 = new ListingStock();
+        stock2.setListingId(10L);
+        stock2.setTicker("IBM2");
+        List<ListingStock> stocks = new ArrayList<>();
+        stocks.add(stock1);
+        stocks.add(stock2);
+        int n = 2;
+
+        ObjectMapper objectMapper1 = new ObjectMapper();
+        listingStockService.setObjectMapper(objectMapper1);
+
+        String response1 = "{\n" +
+                "    \"Meta Data\": {\n" +
+                "        \"1. Information\": \"Daily Prices (open, high, low, close) and Volumes\",\n" +
+                "        \"2. Symbol\": \"IBM\",\n" +
+                "        \"3. Last Refreshed\": \"2024-04-15\",\n" +
+                "        \"4. Output Size\": \"Compact\",\n" +
+                "        \"5. Time Zone\": \"US/Eastern\"\n" +
+                "    },\n" +
+                "    \"Time Series (Daily)\": {\n" +
+                "        \"2024-04-15\": {\n" +
+                "            \"1. open\": \"185.5000\",\n" +
+                "            \"2. high\": \"187.4800\",\n" +
+                "            \"3. low\": \"180.8800\",\n" +
+                "            \"4. close\": \"181.2500\",\n" +
+                "            \"5. volume\": \"3527613\"\n" +
+                "        },\n" +
+                "        \"2024-04-12\": {\n" +
+                "            \"1. open\": \"184.0000\",\n" +
+                "            \"2. high\": \"185.1699\",\n" +
+                "            \"3. low\": \"181.6850\",\n" +
+                "            \"4. close\": \"182.2700\",\n" +
+                "            \"5. volume\": \"3547378\"\n" +
+                "        }}}";
+
+        when(stockRepository.findAll()).thenReturn(stocks);
+        try (MockedStatic<Requests> req = Mockito.mockStatic(Requests.class)) {
+            req.when(() -> Requests.sendRequest(any())).thenReturn(response1);
+
+            List<ListingHistory> listingHistories = listingStockService.fetchNListingsHistory(n);
+
+            assertEquals(4,listingHistories.size());
+        }
+
+    }
+
+
+    @Test
+    public void fetchNListingsHistory_EmptyStockList() {
+        when(stockRepository.findAll()).thenReturn(Collections.emptyList());
+        List<ListingHistory> listingHistories = listingStockService.fetchNListingsHistory(2);
+        assertEquals(0, listingHistories.size());
+    }
+
+
+    @Test
+    public void testGetRequest(){
+        Requests result = new Requests();
+        listingStockService.setRequests(result);
+        assertEquals(listingStockService.getRequests(),result);
+    }
 }
