+++ market-service/src/test/java/rs/edu/raf/banka1/services/ExchangeServiceImplTest.java
@@ -33,7 +33,7 @@ class ExchangeServiceImplTest {
         holidayRepository = mock(HolidayRepository.class);
         ExchangeMapper exchangeMapper = new ExchangeMapper();
 
-        exchangeService = new ExchangeServiceImpl(countryRepository, holidayRepository, exchangeRepository, exchangeMapper);
+        exchangeService = new ExchangeServiceImpl(exchangeRepository, exchangeMapper);
     }
 
     @Test
