+++ market-service/src/test/java/rs/edu/raf/banka1/services/OptionsServiceImplTest.java
@@ -34,212 +34,12 @@ class OptionsServiceImplTest {
     private OptionsServiceImpl optionsService;
     private OptionsRepository optionsRepository;
     private OptionsMapper optionsMapper;
-    private HttpClient httpClientMock;
-    private HttpClient crumbHttpClient;
-    private String cookieMock;
-    private String crumbMock;
-    private String ORCLresponseMockGood = "{\"optionChain\":{\"result\":[{\"underlyingSymbol\":\"ORCL\",\"expirationDates\":[1711065600,1711584000,1712275200,1712880000,1713484800,1714089600,1715904000,1718928000,1721347200,1723766400,1726790400,1734652800,1737072000,1750377600,1766102400,1768521600],\"strikes\":[70.0,75.0,80.0,85.0,90.0,94.0,95.0,96.0,97.0,98.0,99.0,100.0,101.0,102.0,103.0,104.0,105.0,106.0,107.0,108.0,109.0,110.0,111.0,112.0,113.0,114.0,115.0,116.0,117.0,118.0,119.0,120.0,121.0,122.0,123.0,124.0,125.0,126.0,127.0,128.0,129.0,130.0,131.0,132.0,133.0,134.0,135.0,136.0,137.0,138.0,139.0,140.0,141.0,142.0,143.0,144.0,145.0,146.0,147.0,148.0,149.0,150.0,152.5,155.0],\"hasMiniOptions\":false,\"quote\":{\"language\":\"en-US\",\"region\":\"US\",\"quoteType\":\"EQUITY\",\"typeDisp\":\"Equity\",\"quoteSourceName\":\"Nasdaq Real Time Price\",\"triggerable\":true,\"customPriceAlertConfidence\":\"HIGH\",\"currency\":\"USD\",\"marketState\":\"POSTPOST\",\"regularMarketChangePercent\":1.0876364,\"regularMarketPrice\":129.19,\"exchange\":\"NYQ\",\"shortName\":\"Oracle Corporation\",\"longName\":\"Oracle Corporation\",\"messageBoardId\":\"finmb_22247\",\"exchangeTimezoneName\":\"America/New_York\",\"exchangeTimezoneShortName\":\"EDT\",\"gmtOffSetMilliseconds\":-14400000,\"market\":\"us_market\",\"esgPopulated\":false,\"hasPrePostMarketData\":true,\"firstTradeDateMilliseconds\":511021800000,\"priceHint\":2,\"postMarketChangePercent\":-0.18577714,\"postMarketTime\":1710892779,\"postMarketPrice\":128.95,\"postMarketChange\":-0.2400055,\"regularMarketChange\":1.3899994,\"regularMarketTime\":1710878402,\"regularMarketDayHigh\":129.21,\"regularMarketDayRange\":\"126.49 - 129.21\",\"regularMarketDayLow\":126.49,\"regularMarketVolume\":10208162,\"regularMarketPreviousClose\":127.8,\"bid\":129.01,\"ask\":129.25,\"bidSize\":11,\"askSize\":11,\"fullExchangeName\":\"NYSE\",\"financialCurrency\":\"USD\",\"regularMarketOpen\":127.745,\"averageDailyVolume3Month\":8947140,\"averageDailyVolume10Day\":15678780,\"fiftyTwoWeekLowChange\":42.53,\"fiftyTwoWeekLowChangePercent\":0.4907685,\"fiftyTwoWeekRange\":\"86.66 - 129.37\",\"fiftyTwoWeekHighChange\":-0.17999268,\"fiftyTwoWeekHighChangePercent\":-0.0013913015,\"fiftyTwoWeekLow\":86.66,\"fiftyTwoWeekHigh\":129.37,\"fiftyTwoWeekChangePercent\":45.92372,\"dividendDate\":1713916800,\"earningsTimestamp\":1710198000,\"earningsTimestampStart\":1718017140,\"earningsTimestampEnd\":1718366400,\"trailingAnnualDividendRate\":1.6,\"trailingPE\":33.997368,\"dividendRate\":1.6,\"trailingAnnualDividendYield\":0.012519562,\"dividendYield\":1.25,\"epsTrailingTwelveMonths\":3.8,\"epsForward\":6.23,\"epsCurrentYear\":5.59,\"priceEpsCurrentYear\":23.110912,\"sharesOutstanding\":2748509952,\"bookValue\":2.047,\"fiftyDayAverage\":113.0798,\"fiftyDayAverageChange\":16.110199,\"fiftyDayAverageChangePercent\":0.14246751,\"twoHundredDayAverage\":112.79185,\"twoHundredDayAverageChange\":16.398155,\"twoHundredDayAverageChangePercent\":0.14538422,\"marketCap\":355080011776,\"forwardPE\":20.736757,\"priceToBook\":63.111874,\"sourceInterval\":15,\"exchangeDataDelayedBy\":0,\"averageAnalystRating\":\"2.2 - Buy\",\"tradeable\":false,\"cryptoTradeable\":false,\"displayName\":\"Oracle\",\"symbol\":\"ORCL\"},\"options\":[{\"expirationDate\":1711065600,\"hasMiniOptions\":false,\"calls\":[{\"contractSymbol\":\"ORCL240322C00085000\",\"strike\":85.0,\"currency\":\"USD\",\"lastPrice\":42.7,\"change\":0.0,\"percentChange\":0.0,\"volume\":60,\"openInterest\":76,\"bid\":42.7,\"ask\":45.6,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710779107,\"impliedVolatility\":3.480470048828125,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00090000\",\"strike\":90.0,\"currency\":\"USD\",\"lastPrice\":35.85,\"change\":0.0,\"percentChange\":0.0,\"volume\":40,\"openInterest\":66,\"bid\":38.1,\"ask\":40.05,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710439698,\"impliedVolatility\":2.7363312841796876,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00096000\",\"strike\":96.0,\"currency\":\"USD\",\"lastPrice\":32.55,\"change\":13.4,\"percentChange\":69.97389,\"volume\":8,\"openInterest\":23,\"bid\":32.6,\"ask\":34.45,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710872705,\"impliedVolatility\":1.9062504687500001,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00097000\",\"strike\":97.0,\"currency\":\"USD\",\"lastPrice\":31.25,\"change\":13.0,\"percentChange\":71.23288,\"volume\":8,\"openInterest\":8,\"bid\":30.7,\"ask\":33.25,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868796,\"impliedVolatility\":2.3964883837890625,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00098000\",\"strike\":98.0,\"currency\":\"USD\",\"lastPrice\":30.2,\"change\":2.6000004,\"percentChange\":9.420291,\"volume\":8,\"openInterest\":14,\"bid\":30.0,\"ask\":32.35,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868854,\"impliedVolatility\":2.3867227832031244,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00099000\",\"strike\":99.0,\"currency\":\"USD\",\"lastPrice\":29.75,\"change\":3.0599995,\"percentChange\":11.464966,\"volume\":10,\"openInterest\":15,\"bid\":28.3,\"ask\":31.7,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710872248,\"impliedVolatility\":2.5009803100585932,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00100000\",\"strike\":100.0,\"currency\":\"USD\",\"lastPrice\":29.05,\"change\":4.25,\"percentChange\":17.137096,\"volume\":41,\"openInterest\":162,\"bid\":27.9,\"ask\":29.8,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877315,\"impliedVolatility\":1.9150394873046879,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00101000\",\"strike\":101.0,\"currency\":\"USD\",\"lastPrice\":27.5,\"change\":3.0100002,\"percentChange\":12.290731,\"volume\":8,\"openInterest\":26,\"bid\":27.1,\"ask\":29.35,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710869293,\"impliedVolatility\":1.140629296875,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00102000\",\"strike\":102.0,\"currency\":\"USD\",\"lastPrice\":26.35,\"change\":2.8500004,\"percentChange\":12.127661,\"volume\":8,\"openInterest\":33,\"bid\":26.4,\"ask\":27.65,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868881,\"impliedVolatility\":1.6845718896484376,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00103000\",\"strike\":103.0,\"currency\":\"USD\",\"lastPrice\":25.25,\"change\":0.6100006,\"percentChange\":2.475652,\"volume\":8,\"openInterest\":48,\"bid\":25.45,\"ask\":26.65,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868601,\"impliedVolatility\":1.6289081054687498,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00104000\",\"strike\":104.0,\"currency\":\"USD\",\"lastPrice\":24.95,\"change\":0.5799999,\"percentChange\":2.3799748,\"volume\":8,\"openInterest\":28,\"bid\":24.8,\"ask\":25.55,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710873151,\"impliedVolatility\":1.4912134814453122,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00105000\",\"strike\":105.0,\"currency\":\"USD\",\"lastPrice\":23.8,\"change\":0.8999996,\"percentChange\":3.9301295,\"volume\":11,\"openInterest\":62,\"bid\":23.9,\"ask\":25.15,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710872960,\"impliedVolatility\":1.4160185449218752,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00106000\",\"strike\":106.0,\"currency\":\"USD\",\"lastPrice\":22.5,\"change\":2.1499996,\"percentChange\":10.565109,\"volume\":8,\"openInterest\":35,\"bid\":22.05,\"ask\":24.1,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710869495,\"impliedVolatility\":1.732423212890625,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00107000\",\"strike\":107.0,\"currency\":\"USD\",\"lastPrice\":21.95,\"change\":2.1500015,\"percentChange\":10.858594,\"volume\":8,\"openInterest\":123,\"bid\":21.0,\"ask\":22.7,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710870081,\"impliedVolatility\":1.439455927734375,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00108000\",\"strike\":108.0,\"currency\":\"USD\",\"lastPrice\":21.0,\"change\":1.0,\"percentChange\":5.0,\"volume\":10,\"openInterest\":71,\"bid\":20.0,\"ask\":22.0,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710873552,\"impliedVolatility\":1.5576193994140621,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00109000\",\"strike\":109.0,\"currency\":\"USD\",\"lastPrice\":19.7,\"change\":0.8200016,\"percentChange\":4.3432293,\"volume\":8,\"openInterest\":111,\"bid\":19.05,\"ask\":21.1,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710872424,\"impliedVolatility\":1.5468772656249998,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00110000\",\"strike\":110.0,\"currency\":\"USD\",\"lastPrice\":18.85,\"change\":1.3500004,\"percentChange\":7.7142878,\"volume\":11,\"openInterest\":202,\"bid\":18.95,\"ask\":19.75,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710873060,\"impliedVolatility\":0.9941406835937501,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00111000\",\"strike\":111.0,\"currency\":\"USD\",\"lastPrice\":17.86,\"change\":3.1200008,\"percentChange\":21.166899,\"volume\":14,\"openInterest\":297,\"bid\":17.4,\"ask\":18.7,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710876007,\"impliedVolatility\":1.2168007910156249,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00112000\",\"strike\":112.0,\"currency\":\"USD\",\"lastPrice\":17.21,\"change\":1.0099983,\"percentChange\":6.234557,\"volume\":23,\"openInterest\":421,\"bid\":16.95,\"ask\":17.75,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877707,\"impliedVolatility\":0.9023447265625,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00113000\",\"strike\":113.0,\"currency\":\"USD\",\"lastPrice\":15.89,\"change\":0.96000004,\"percentChange\":6.430007,\"volume\":2,\"openInterest\":216,\"bid\":16.0,\"ask\":17.3,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710876726,\"impliedVolatility\":1.0771530517578127,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00114000\",\"strike\":114.0,\"currency\":\"USD\",\"lastPrice\":14.65,\"change\":1.1999998,\"percentChange\":8.921931,\"volume\":8,\"openInterest\":1333,\"bid\":14.0,\"ask\":16.1,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710867287,\"impliedVolatility\":1.2392616162109373,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00115000\",\"strike\":115.0,\"currency\":\"USD\",\"lastPrice\":13.94,\"change\":0.9199991,\"percentChange\":7.066045,\"volume\":8,\"openInterest\":751,\"bid\":14.0,\"ask\":15.05,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710873520,\"impliedVolatility\":0.8945323046874999,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00116000\",\"strike\":116.0,\"currency\":\"USD\",\"lastPrice\":12.2,\"change\":0.5,\"percentChange\":4.2735043,\"volume\":52,\"openInterest\":450,\"bid\":13.0,\"ask\":14.1,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868820,\"impliedVolatility\":0.85742330078125,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00117000\",\"strike\":117.0,\"currency\":\"USD\",\"lastPrice\":11.26,\"change\":0.11000061,\"percentChange\":0.9865526,\"volume\":13,\"openInterest\":259,\"bid\":12.05,\"ask\":12.95,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868672,\"impliedVolatility\":0.7753928710937501,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00118000\",\"strike\":118.0,\"currency\":\"USD\",\"lastPrice\":11.15,\"change\":1.0999994,\"percentChange\":10.945268,\"volume\":11,\"openInterest\":830,\"bid\":11.0,\"ask\":11.7,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877421,\"impliedVolatility\":0.62305064453125,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00119000\",\"strike\":119.0,\"currency\":\"USD\",\"lastPrice\":10.2,\"change\":1.1099997,\"percentChange\":12.211217,\"volume\":7,\"openInterest\":471,\"bid\":9.1,\"ask\":10.45,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878372,\"impliedVolatility\":0.6435582519531251,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00120000\",\"strike\":120.0,\"currency\":\"USD\",\"lastPrice\":9.2,\"change\":0.80999947,\"percentChange\":9.654344,\"volume\":48,\"openInterest\":1267,\"bid\":8.85,\"ask\":9.35,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878370,\"impliedVolatility\":0.5293015820312501,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00121000\",\"strike\":121.0,\"currency\":\"USD\",\"lastPrice\":7.88,\"change\":0.98,\"percentChange\":14.202899,\"volume\":8,\"openInterest\":1396,\"bid\":8.0,\"ask\":8.55,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710875564,\"impliedVolatility\":0.5918009570312501,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00122000\",\"strike\":122.0,\"currency\":\"USD\",\"lastPrice\":7.0,\"change\":1.0,\"percentChange\":16.666668,\"volume\":20,\"openInterest\":277,\"bid\":6.2,\"ask\":7.95,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710876883,\"impliedVolatility\":0.6875031250000001,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00123000\",\"strike\":123.0,\"currency\":\"USD\",\"lastPrice\":5.44,\"change\":0.13999987,\"percentChange\":2.641507,\"volume\":1,\"openInterest\":347,\"bid\":6.0,\"ask\":7.05,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710866379,\"impliedVolatility\":0.6557651611328126,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00124000\",\"strike\":124.0,\"currency\":\"USD\",\"lastPrice\":5.28,\"change\":0.98,\"percentChange\":22.790697,\"volume\":47,\"openInterest\":2562,\"bid\":5.05,\"ask\":5.5,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877891,\"impliedVolatility\":0.405279384765625,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00125000\",\"strike\":125.0,\"currency\":\"USD\",\"lastPrice\":4.35,\"change\":0.6999998,\"percentChange\":19.178076,\"volume\":2506,\"openInterest\":4236,\"bid\":4.25,\"ask\":4.55,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878384,\"impliedVolatility\":0.3676821044921875,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00126000\",\"strike\":126.0,\"currency\":\"USD\",\"lastPrice\":3.3,\"change\":0.43000007,\"percentChange\":14.982581,\"volume\":862,\"openInterest\":2598,\"bid\":3.15,\"ask\":3.55,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878363,\"impliedVolatility\":0.30762411132812495,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00127000\",\"strike\":127.0,\"currency\":\"USD\",\"lastPrice\":2.7,\"change\":0.51,\"percentChange\":23.28767,\"volume\":3086,\"openInterest\":4815,\"bid\":2.57,\"ask\":2.83,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878398,\"impliedVolatility\":0.32031929687499994,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00128000\",\"strike\":128.0,\"currency\":\"USD\",\"lastPrice\":1.95,\"change\":0.24000001,\"percentChange\":14.035088,\"volume\":7435,\"openInterest\":4848,\"bid\":1.95,\"ask\":1.99,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878397,\"impliedVolatility\":0.28174546386718746,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00129000\",\"strike\":129.0,\"currency\":\"USD\",\"lastPrice\":1.27,\"change\":-0.00999999,\"percentChange\":-0.7812493,\"volume\":2074,\"openInterest\":1782,\"bid\":1.37,\"ask\":1.4,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878365,\"impliedVolatility\":0.2788158056640625,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322C00130000\",\"strike\":130.0,\"currency\":\"USD\",\"lastPrice\":0.91,\"change\":-0.089999974,\"percentChange\":-8.999997,\"volume\":7573,\"openInterest\":4371,\"bid\":0.91,\"ask\":0.94,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878398,\"impliedVolatility\":0.278327529296875,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00131000\",\"strike\":131.0,\"currency\":\"USD\",\"lastPrice\":0.52,\"change\":-0.22000003,\"percentChange\":-29.729733,\"volume\":809,\"openInterest\":10373,\"bid\":0.57,\"ask\":0.63,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878278,\"impliedVolatility\":0.28613995117187496,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00132000\",\"strike\":132.0,\"currency\":\"USD\",\"lastPrice\":0.36,\"change\":-0.18,\"percentChange\":-33.333336,\"volume\":444,\"openInterest\":1773,\"bid\":0.33,\"ask\":0.38,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878102,\"impliedVolatility\":0.2841868457031249,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00133000\",\"strike\":133.0,\"currency\":\"USD\",\"lastPrice\":0.2,\"change\":-0.21999998,\"percentChange\":-52.38095,\"volume\":386,\"openInterest\":1663,\"bid\":0.19,\"ask\":0.24,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878337,\"impliedVolatility\":0.2919992675781249,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00134000\",\"strike\":134.0,\"currency\":\"USD\",\"lastPrice\":0.13,\"change\":-0.18,\"percentChange\":-58.06452,\"volume\":97,\"openInterest\":1476,\"bid\":0.12,\"ask\":0.15,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877992,\"impliedVolatility\":0.3007882421875,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00135000\",\"strike\":135.0,\"currency\":\"USD\",\"lastPrice\":0.09,\"change\":-0.16,\"percentChange\":-64.0,\"volume\":1312,\"openInterest\":3210,\"bid\":0.07,\"ask\":0.1,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878257,\"impliedVolatility\":0.31445998046874996,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00136000\",\"strike\":136.0,\"currency\":\"USD\",\"lastPrice\":0.07,\"change\":-0.12,\"percentChange\":-63.15789,\"volume\":111,\"openInterest\":121,\"bid\":0.05,\"ask\":0.07,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878027,\"impliedVolatility\":0.33008482421874996,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00137000\",\"strike\":137.0,\"currency\":\"USD\",\"lastPrice\":0.04,\"change\":-0.1,\"percentChange\":-71.42857,\"volume\":15,\"openInterest\":175,\"bid\":0.04,\"ask\":0.05,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877371,\"impliedVolatility\":0.34570966796874997,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00138000\",\"strike\":138.0,\"currency\":\"USD\",\"lastPrice\":0.04,\"change\":-0.1,\"percentChange\":-71.42857,\"volume\":79,\"openInterest\":309,\"bid\":0.03,\"ask\":0.04,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878217,\"impliedVolatility\":0.367193828125,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00139000\",\"strike\":139.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":-0.07,\"percentChange\":-70.0,\"volume\":1,\"openInterest\":63,\"bid\":0.02,\"ask\":0.04,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868385,\"impliedVolatility\":0.4023497265625,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00140000\",\"strike\":140.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":-0.060000002,\"percentChange\":-66.66667,\"volume\":154,\"openInterest\":1325,\"bid\":0.02,\"ask\":0.04,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878035,\"impliedVolatility\":0.4335994140625,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00141000\",\"strike\":141.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":-0.049999997,\"percentChange\":-62.5,\"volume\":1,\"openInterest\":10,\"bid\":0.01,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710865722,\"impliedVolatility\":0.445318046875,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00142000\",\"strike\":142.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.080000006,\"percentChange\":-88.88889,\"volume\":2,\"openInterest\":12,\"bid\":0.01,\"ask\":0.07,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710856015,\"impliedVolatility\":0.5429733203125001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00143000\",\"strike\":143.0,\"currency\":\"USD\",\"lastPrice\":0.08,\"change\":0.0,\"percentChange\":0.0,\"volume\":17,\"openInterest\":56,\"bid\":0.01,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710777910,\"impliedVolatility\":0.4804739453125,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00144000\",\"strike\":144.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":-0.020000001,\"percentChange\":-40.000004,\"volume\":1,\"openInterest\":154,\"bid\":0.01,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710856730,\"impliedVolatility\":0.507817421875,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00145000\",\"strike\":145.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.04,\"percentChange\":-80.0,\"volume\":3,\"openInterest\":767,\"bid\":0.01,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868033,\"impliedVolatility\":0.5390671093750001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00146000\",\"strike\":146.0,\"currency\":\"USD\",\"lastPrice\":0.05,\"change\":0.0,\"percentChange\":0.0,\"volume\":10,\"openInterest\":15,\"bid\":0.0,\"ask\":0.25,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710516524,\"impliedVolatility\":0.7402369726562499,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00147000\",\"strike\":147.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":0.0,\"percentChange\":0.0,\"volume\":12,\"openInterest\":13,\"bid\":0.0,\"ask\":0.04,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710790435,\"impliedVolatility\":0.5937540625000001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00148000\",\"strike\":148.0,\"currency\":\"USD\",\"lastPrice\":0.02,\"change\":-0.02,\"percentChange\":-50.0,\"volume\":2,\"openInterest\":101,\"bid\":0.01,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877314,\"impliedVolatility\":0.6171913281250001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00149000\",\"strike\":149.0,\"currency\":\"USD\",\"lastPrice\":0.14,\"change\":0.0,\"percentChange\":0.0,\"openInterest\":3,\"bid\":0.0,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710264147,\"impliedVolatility\":0.6250037500000001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00150000\",\"strike\":150.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.02,\"percentChange\":-66.66667,\"volume\":138,\"openInterest\":694,\"bid\":0.01,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710871236,\"impliedVolatility\":0.6718782812500002,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00152500\",\"strike\":152.5,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.04,\"percentChange\":-80.0,\"volume\":113,\"openInterest\":108,\"bid\":0.01,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710875779,\"impliedVolatility\":0.7187528125,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322C00155000\",\"strike\":155.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.01,\"percentChange\":-50.0,\"volume\":6,\"openInterest\":1230,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710866833,\"impliedVolatility\":0.7500025,\"inTheMoney\":false}],\"puts\":[{\"contractSymbol\":\"ORCL240322P00070000\",\"strike\":70.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"openInterest\":4,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710187109,\"impliedVolatility\":2.1875045312499997,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00075000\",\"strike\":75.0,\"currency\":\"USD\",\"lastPrice\":0.02,\"change\":0.0,\"percentChange\":0.0,\"openInterest\":40,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710186424,\"impliedVolatility\":1.9375003125,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00080000\",\"strike\":80.0,\"currency\":\"USD\",\"lastPrice\":0.04,\"change\":0.0,\"percentChange\":0.0,\"volume\":585,\"openInterest\":530,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710187169,\"impliedVolatility\":1.75000125,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00085000\",\"strike\":85.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":95,\"openInterest\":310,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710265367,\"impliedVolatility\":1.5000025,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00090000\",\"strike\":90.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":2,\"openInterest\":997,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710768632,\"impliedVolatility\":1.3125034374999998,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00094000\",\"strike\":94.0,\"currency\":\"USD\",\"lastPrice\":0.02,\"change\":0.0,\"percentChange\":0.0,\"volume\":31,\"openInterest\":315,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710263966,\"impliedVolatility\":1.2500037499999999,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00095000\",\"strike\":95.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":2,\"openInterest\":382,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710442344,\"impliedVolatility\":1.125004375,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00096000\",\"strike\":96.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.01,\"percentChange\":-50.0,\"volume\":7,\"openInterest\":180,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710872765,\"impliedVolatility\":1.1875040625,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00097000\",\"strike\":97.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":7,\"openInterest\":87,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710861648,\"impliedVolatility\":1.0625046875000002,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00098000\",\"strike\":98.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":6,\"openInterest\":333,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710521469,\"impliedVolatility\":1.0312548437500002,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00099000\",\"strike\":99.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":7,\"openInterest\":313,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710855330,\"impliedVolatility\":1.0625046875000002,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00100000\",\"strike\":100.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":0.0,\"percentChange\":0.0,\"volume\":1,\"openInterest\":5078,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710769002,\"impliedVolatility\":0.9687503125,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00101000\",\"strike\":101.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":4,\"openInterest\":124,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710863459,\"impliedVolatility\":0.992187578125,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00102000\",\"strike\":102.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":5,\"openInterest\":267,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710774890,\"impliedVolatility\":0.9531254687499999,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00103000\",\"strike\":103.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":52,\"openInterest\":382,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710781109,\"impliedVolatility\":0.87500125,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00104000\",\"strike\":104.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":4,\"openInterest\":549,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710529073,\"impliedVolatility\":0.8906260937499999,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00105000\",\"strike\":105.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":161,\"openInterest\":1281,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710791804,\"impliedVolatility\":0.8437515624999999,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00106000\",\"strike\":106.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":0.0,\"percentChange\":0.0,\"volume\":1,\"openInterest\":605,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710530741,\"impliedVolatility\":0.8125018749999999,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00107000\",\"strike\":107.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":0.02,\"percentChange\":200.0,\"volume\":2,\"openInterest\":249,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710876077,\"impliedVolatility\":0.7812521875,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00108000\",\"strike\":108.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":10,\"openInterest\":404,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710867405,\"impliedVolatility\":0.6875031250000001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00109000\",\"strike\":109.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.02,\"percentChange\":-66.66667,\"volume\":88,\"openInterest\":360,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868040,\"impliedVolatility\":0.6562534375000001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00110000\",\"strike\":110.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.01,\"percentChange\":-50.0,\"volume\":5,\"openInterest\":2382,\"bid\":0.0,\"ask\":0.01,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710869811,\"impliedVolatility\":0.6250037500000001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00111000\",\"strike\":111.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":0.0,\"percentChange\":0.0,\"volume\":5,\"openInterest\":231,\"bid\":0.0,\"ask\":0.23,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710856764,\"impliedVolatility\":0.8945323046874999,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00112000\",\"strike\":112.0,\"currency\":\"USD\",\"lastPrice\":0.01,\"change\":-0.01,\"percentChange\":-50.0,\"volume\":31,\"openInterest\":767,\"bid\":0.0,\"ask\":0.05,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710868129,\"impliedVolatility\":0.6796907031250001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00113000\",\"strike\":113.0,\"currency\":\"USD\",\"lastPrice\":0.02,\"change\":0.0,\"percentChange\":0.0,\"volume\":4,\"openInterest\":739,\"bid\":0.01,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710791944,\"impliedVolatility\":0.6015664843750002,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00114000\",\"strike\":114.0,\"currency\":\"USD\",\"lastPrice\":0.02,\"change\":-0.02,\"percentChange\":-50.0,\"volume\":59,\"openInterest\":771,\"bid\":0.0,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710874995,\"impliedVolatility\":0.5390671093750001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00115000\",\"strike\":115.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":0.01,\"percentChange\":50.0,\"volume\":16,\"openInterest\":336,\"bid\":0.02,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710857439,\"impliedVolatility\":0.5625043750000001,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00116000\",\"strike\":116.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":0.0,\"percentChange\":0.0,\"volume\":11,\"openInterest\":690,\"bid\":0.01,\"ask\":0.02,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710785416,\"impliedVolatility\":0.51562984375,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00117000\",\"strike\":117.0,\"currency\":\"USD\",\"lastPrice\":0.06,\"change\":0.02,\"percentChange\":50.0,\"volume\":3,\"openInterest\":1471,\"bid\":0.01,\"ask\":0.05,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710876077,\"impliedVolatility\":0.5039112109375,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00118000\",\"strike\":118.0,\"currency\":\"USD\",\"lastPrice\":0.02,\"change\":-0.030000001,\"percentChange\":-60.000004,\"volume\":71,\"openInterest\":478,\"bid\":0.02,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710871938,\"impliedVolatility\":0.46484910156250003,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00119000\",\"strike\":119.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":-0.07,\"percentChange\":-70.0,\"volume\":11,\"openInterest\":288,\"bid\":0.02,\"ask\":0.03,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710866146,\"impliedVolatility\":0.42969320312500003,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00120000\",\"strike\":120.0,\"currency\":\"USD\",\"lastPrice\":0.03,\"change\":-0.089999996,\"percentChange\":-75.0,\"volume\":61,\"openInterest\":643,\"bid\":0.03,\"ask\":0.04,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710865577,\"impliedVolatility\":0.4101621484375,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00121000\",\"strike\":121.0,\"currency\":\"USD\",\"lastPrice\":0.05,\"change\":-0.10000001,\"percentChange\":-66.66667,\"volume\":36,\"openInterest\":129,\"bid\":0.03,\"ask\":0.04,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877281,\"impliedVolatility\":0.36914693359375,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00122000\",\"strike\":122.0,\"currency\":\"USD\",\"lastPrice\":0.04,\"change\":-0.19999999,\"percentChange\":-83.333336,\"volume\":294,\"openInterest\":547,\"bid\":0.04,\"ask\":0.06,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877899,\"impliedVolatility\":0.35352208984374994,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00123000\",\"strike\":123.0,\"currency\":\"USD\",\"lastPrice\":0.09,\"change\":-0.28,\"percentChange\":-75.675674,\"volume\":112,\"openInterest\":686,\"bid\":0.06,\"ask\":0.08,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710877002,\"impliedVolatility\":0.33008482421874996,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00124000\",\"strike\":124.0,\"currency\":\"USD\",\"lastPrice\":0.12,\"change\":-0.39999998,\"percentChange\":-76.92307,\"volume\":630,\"openInterest\":803,\"bid\":0.1,\"ask\":0.12,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878000,\"impliedVolatility\":0.313483427734375,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00125000\",\"strike\":125.0,\"currency\":\"USD\",\"lastPrice\":0.17,\"change\":-0.59,\"percentChange\":-77.63158,\"volume\":761,\"openInterest\":2530,\"bid\":0.15,\"ask\":0.19,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878378,\"impliedVolatility\":0.3007882421875,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00126000\",\"strike\":126.0,\"currency\":\"USD\",\"lastPrice\":0.28,\"change\":-0.78,\"percentChange\":-73.58491,\"volume\":727,\"openInterest\":1525,\"bid\":0.26,\"ask\":0.28,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878399,\"impliedVolatility\":0.28076891113281244,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00127000\",\"strike\":127.0,\"currency\":\"USD\",\"lastPrice\":0.46,\"change\":-0.93999994,\"percentChange\":-67.14285,\"volume\":946,\"openInterest\":1356,\"bid\":0.42,\"ask\":0.45,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878395,\"impliedVolatility\":0.270515107421875,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00128000\",\"strike\":128.0,\"currency\":\"USD\",\"lastPrice\":0.72,\"change\":-1.18,\"percentChange\":-62.105263,\"volume\":2296,\"openInterest\":795,\"bid\":0.68,\"ask\":0.72,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878397,\"impliedVolatility\":0.26319096191406255,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00129000\",\"strike\":129.0,\"currency\":\"USD\",\"lastPrice\":1.2,\"change\":-1.3399999,\"percentChange\":-52.755905,\"volume\":333,\"openInterest\":297,\"bid\":1.08,\"ask\":1.12,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878365,\"impliedVolatility\":0.25977302734375,\"inTheMoney\":false},{\"contractSymbol\":\"ORCL240322P00130000\",\"strike\":130.0,\"currency\":\"USD\",\"lastPrice\":1.78,\"change\":-1.3700001,\"percentChange\":-43.492065,\"volume\":3369,\"openInterest\":2311,\"bid\":1.61,\"ask\":1.66,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710878365,\"impliedVolatility\":0.25879647460937505,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00131000\",\"strike\":131.0,\"currency\":\"USD\",\"lastPrice\":3.8,\"change\":0.0,\"percentChange\":0.0,\"volume\":51,\"openInterest\":94,\"bid\":2.26,\"ask\":2.52,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710787108,\"impliedVolatility\":0.305671005859375,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00132000\",\"strike\":132.0,\"currency\":\"USD\",\"lastPrice\":3.6,\"change\":-1.3000002,\"percentChange\":-26.530615,\"volume\":7,\"openInterest\":46,\"bid\":2.82,\"ask\":3.25,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710867042,\"impliedVolatility\":0.301764794921875,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00133000\",\"strike\":133.0,\"currency\":\"USD\",\"lastPrice\":5.45,\"change\":0.0,\"percentChange\":0.0,\"volume\":36,\"openInterest\":53,\"bid\":3.8,\"ask\":4.15,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710787188,\"impliedVolatility\":0.32813171874999997,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00134000\",\"strike\":134.0,\"currency\":\"USD\",\"lastPrice\":6.25,\"change\":0.0,\"percentChange\":0.0,\"volume\":63,\"openInterest\":68,\"bid\":4.1,\"ask\":6.15,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710785867,\"impliedVolatility\":0.6689486230468751,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00135000\",\"strike\":135.0,\"currency\":\"USD\",\"lastPrice\":7.4,\"change\":0.0,\"percentChange\":0.0,\"volume\":9,\"openInterest\":31,\"bid\":5.15,\"ask\":6.1,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710783245,\"impliedVolatility\":0.41211525390625,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00136000\",\"strike\":136.0,\"currency\":\"USD\",\"lastPrice\":11.1,\"change\":0.0,\"percentChange\":0.0,\"openInterest\":56,\"bid\":5.85,\"ask\":8.05,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710251340,\"impliedVolatility\":0.7690452783203126,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00137000\",\"strike\":137.0,\"currency\":\"USD\",\"lastPrice\":11.6,\"change\":0.0,\"percentChange\":0.0,\"openInterest\":15,\"bid\":7.15,\"ask\":9.0,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710359400,\"impliedVolatility\":0.81445498046875,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00138000\",\"strike\":138.0,\"currency\":\"USD\",\"lastPrice\":10.6,\"change\":0.0,\"percentChange\":0.0,\"volume\":27,\"openInterest\":35,\"bid\":8.2,\"ask\":10.1,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710773971,\"impliedVolatility\":0.5747112841796875,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00140000\",\"strike\":140.0,\"currency\":\"USD\",\"lastPrice\":13.0,\"change\":1.0,\"percentChange\":8.333334,\"volume\":7,\"openInterest\":7,\"bid\":10.45,\"ask\":11.0,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710855767,\"impliedVolatility\":0.577152666015625,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00141000\",\"strike\":141.0,\"currency\":\"USD\",\"lastPrice\":15.9,\"change\":0.0,\"percentChange\":0.0,\"openInterest\":7,\"bid\":11.0,\"ask\":12.25,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710350415,\"impliedVolatility\":0.754885263671875,\"inTheMoney\":true},{\"contractSymbol\":\"ORCL240322P00149000\",\"strike\":149.0,\"currency\":\"USD\",\"lastPrice\":21.9,\"change\":0.0,\"percentChange\":0.0,\"openInterest\":10,\"bid\":19.05,\"ask\":21.95,\"contractSize\":\"REGULAR\",\"expiration\":1711065600,\"lastTradeDate\":1710273519,\"impliedVolatility\":1.2089883300781248,\"inTheMoney\":true}]}]}],\"error\":null}}";
-    private String NEXOYresponseOptionsEmpty = "{\"optionChain\":{\"result\":[{\"underlyingSymbol\":\"NEXOY\",\"expirationDates\":[],\"strikes\":[],\"hasMiniOptions\":false,\"quote\":{\"language\":\"en-US\",\"region\":\"US\",\"quoteType\":\"EQUITY\",\"typeDisp\":\"Equity\",\"quoteSourceName\":\"Delayed Quote\",\"triggerable\":false,\"customPriceAlertConfidence\":\"LOW\",\"currency\":\"USD\",\"marketState\":\"POST\",\"regularMarketChangePercent\":0.23242834,\"regularMarketPrice\":17.25,\"exchange\":\"PNK\",\"shortName\":\"NEXON CO LTD UNSP ADR EACH REPR\",\"longName\":\"NEXON Co., Ltd.\",\"messageBoardId\":\"finmb_142745502\",\"exchangeTimezoneName\":\"America/New_York\",\"exchangeTimezoneShortName\":\"EDT\",\"gmtOffSetMilliseconds\":-14400000,\"market\":\"us_market\",\"esgPopulated\":false,\"hasPrePostMarketData\":false,\"firstTradeDateMilliseconds\":1525786200000,\"priceHint\":2,\"regularMarketChange\":0.040000916,\"regularMarketTime\":1710878399,\"regularMarketDayHigh\":17.25,\"regularMarketDayRange\":\"17.1075 - 17.25\",\"regularMarketDayLow\":17.1075,\"regularMarketVolume\":46174,\"regularMarketPreviousClose\":17.21,\"fullExchangeName\":\"Other OTC\",\"financialCurrency\":\"JPY\",\"regularMarketOpen\":17.125,\"averageDailyVolume3Month\":62290,\"averageDailyVolume10Day\":133280,\"fiftyTwoWeekLowChange\":1.9499998,\"fiftyTwoWeekLowChangePercent\":0.12745097,\"fiftyTwoWeekRange\":\"15.3 - 24.39\",\"fiftyTwoWeekHighChange\":-7.1399994,\"fiftyTwoWeekHighChangePercent\":-0.2927429,\"fiftyTwoWeekLow\":15.3,\"fiftyTwoWeekHigh\":24.39,\"fiftyTwoWeekChangePercent\":-24.978203,\"dividendDate\":1586390400,\"trailingAnnualDividendRate\":0.0,\"trailingPE\":21.036585,\"dividendRate\":0.07,\"trailingAnnualDividendYield\":0.0,\"dividendYield\":0.42,\"epsTrailingTwelveMonths\":0.82,\"sharesOutstanding\":840278016,\"bookValue\":1059.179,\"fiftyDayAverage\":17.17808,\"fiftyDayAverageChange\":0.071920395,\"fiftyDayAverageChangePercent\":0.004186754,\"twoHundredDayAverage\":18.766,\"twoHundredDayAverageChange\":-1.5160007,\"twoHundredDayAverageChangePercent\":-0.08078443,\"marketCap\":14530295808,\"priceToBook\":0.0162862,\"sourceInterval\":15,\"exchangeDataDelayedBy\":0,\"tradeable\":false,\"cryptoTradeable\":false,\"displayName\":\"NEXON\",\"symbol\":\"NEXOY\"},\"options\":[]}],\"error\":null}}";
-    private String GMGIDresponseResultEmpty = "{\"optionChain\":{\"result\":[],\"error\":null}}";
-    private String NLOK404error = "{\"finance\":{\"result\":null,\"error\":{\"code\":\"Not Found\",\"description\":\"Missing OptionType for {contract: NLOK250117C00032000 }, {tickerSymbol: NLOK }, {strike: 32.0 }, {expirationDate: 1737072000 }\"}}}";
-    private HttpRequest httpRequestMock;
-    private HttpResponse responseMock;
-    private HttpResponse crumbResponseMock;
-
 
     @BeforeEach
     public void setUp() {
         optionsRepository = mock(OptionsRepository.class);
-        httpClientMock = mock(HttpClient.class);
-        crumbHttpClient = mock(HttpClient.class);
-        httpRequestMock = mock(HttpRequest.class);
-        responseMock = mock(HttpResponse.class);
-        crumbResponseMock = mock(HttpResponse.class);
-        crumbMock = "crumbMock";
-        cookieMock = "cookieMock";
         optionsMapper = new OptionsMapper();
         optionsService = new OptionsServiceImpl(optionsRepository, optionsMapper);
-        optionsService.setHttpClient(httpClientMock);
-        optionsService.setCrumbHttpClient(crumbHttpClient);
-        optionsService.setHttpRequest(httpRequestMock);
-        optionsService.setCrumb(crumbMock);
-        optionsService.setCookie(cookieMock);
-    }
-
-    @Test
-    public void testFetchOptionsByTickerHappy() throws IOException, InterruptedException {
-//        when(exchangeRepository.findById(id)).thenReturn(Optional.of(exchange));
-        // Mock the HTTP response
-        when(responseMock.statusCode()).thenReturn(200);
-        when(responseMock.body()).thenReturn(ORCLresponseMockGood);
-        // Mock the HttpClient behavior
-        when(httpClientMock.send(Mockito.any(), Mockito.any())).thenReturn(responseMock);
-        // Invoke the method to test
-        String url = "https://query2.finance.yahoo.com/v7/finance/options/ORCL";
-        List<OptionsDto> options = optionsService.fetchOptionsForTicker("ORCL", url);
-
-        // Verify the result
-        assertNotNull(options);
-        Assertions.assertFalse(options.isEmpty());
-    }
-
-    @Test
-    public void testFetchOptionsByTickerNoOptions() throws IOException, InterruptedException {
-        when(responseMock.statusCode()).thenReturn(200);
-        when(responseMock.body()).thenReturn(NEXOYresponseOptionsEmpty);
-        when(httpClientMock.send(Mockito.any(), Mockito.any())).thenReturn(responseMock);
-        // Invoke the method to test
-        String url = "https://query2.finance.yahoo.com/v7/finance/options/NEXOY";
-        List<OptionsDto> options = optionsService.fetchOptionsForTicker("NEXOY", url);
-
-        // Verify the result
-        assertNotNull(options);
-        assertTrue(options.isEmpty());
-    }
-
-    @Test
-    public void testFetchOptionsByTickerNoResult() throws IOException, InterruptedException {
-        // Mock the HTTP response
-        when(responseMock.statusCode()).thenReturn(200);
-        when(responseMock.body()).thenReturn(GMGIDresponseResultEmpty);
-        // Mock the HttpClient behavior
-        when(httpClientMock.send(Mockito.any(), Mockito.any())).thenReturn(responseMock);
-        // Invoke the method to test
-        String url = "https://query2.finance.yahoo.com/v7/finance/options/GMGID";
-        List<OptionsDto> options = optionsService.fetchOptionsForTicker("GMGID", url);
-
-        // Verify the result
-        assertNotNull(options);
-        assertTrue(options.isEmpty());
-    }
-
-    @Test
-    public void testFetchOptionsByTickerErrorCode() throws IOException, InterruptedException {
-        // Mock the HTTP response
-        when(responseMock.statusCode()).thenReturn(200);
-        when(responseMock.body()).thenReturn(NLOK404error);
-        // Mock the HttpClient behavior
-        when(httpClientMock.send(Mockito.any(), Mockito.any())).thenReturn(responseMock);
-        // Invoke the method to test
-        String url = "https://query2.finance.yahoo.com/v7/finance/options/NLOK";
-        List<OptionsDto> options = optionsService.fetchOptionsForTicker("NLOK", url);
-
-        // Verify the result
-        assertNotNull(options);
-        assertTrue(options.isEmpty());
-//        Assertions.assertThrows(OptionsException.class, () -> optionsService.fetchOptionsForTicker("NLOK", url));
-    }
-
-    ////////////////////////////////////////////////////////////////////////////////////////
-    @Test
-    public void testNonValidCookieFalse() throws IOException, InterruptedException {
-        when(responseMock.statusCode()).thenReturn(404);
-        when(crumbResponseMock.statusCode()).thenReturn(429);
-        // Create a map with sample data
-        Map<String, List<String>> headersMap = new HashMap<>();
-        headersMap.put("Set-Cookie", List.of("nonValidCookie"));
-        HttpHeaders headersMock = mock(HttpHeaders.class);
-        when(responseMock.headers()).thenReturn(headersMock);
-        when(responseMock.headers().map()).thenReturn(headersMap);
-
-        when(httpClientMock.send(any(), any())).thenReturn(responseMock);
-        when(crumbHttpClient.send(any(), any())).thenReturn(crumbResponseMock);
-//        when(optionsService.getCookieAndCrumb()).thenReturn(false); // non valid Cookie vraca false
-        optionsService.setCookie(null);
-        optionsService.setCrumb(null);
-        List<OptionsDto> options = optionsService.fetchOptions();
-
-        assertNotNull(options);
-        assertTrue(options.isEmpty());
-    }
- // ako je non-valid crumb treba da vrati 401 ali fetchByTicker !!!
-    @Test
-    public void testNonValidCrumbFalse() throws IOException, InterruptedException {
-        when(responseMock.statusCode()).thenReturn(401);
-        // Mock the HttpClient behavior
-        when(httpClientMock.send(Mockito.any(), Mockito.any())).thenReturn(responseMock);
-        // Invoke the method to test
-        String url = "https://query2.finance.yahoo.com/v7/finance/options/ORCL?crumb=badCrumb";
-        List<OptionsDto> options = optionsService.fetchOptionsForTicker("ORCL", url);
-
-        // Verify the result
-        assertNotNull(options);
-        assertTrue(options.isEmpty());
-    }
-
-    @Test
-    public void truncateTableTest() {
-        optionsService.truncateTable();
-        verify(optionsRepository, times(1)).truncateTable();
-    }
-
-    @Test
-    public void testTruncateAndFetch() {
-        // Mock the behavior of optionsRepository.truncateTable() and optionsRepository.fetchOptions()
-        Mockito.doNothing().when(optionsRepository).truncateTable();
-        // Call the method to be tested
-        optionsService.truncateAndFetch();
-        // Verify that the methods were called once
-        verify(optionsRepository, times(1)).truncateTable();
-    }
-
-    @Test
-    public void fetchTickersTest_Success(){
-        JsonNode rootNode = mock(JsonNode.class);
-        JsonNode node1 = mock(JsonNode.class);
-        JsonNode node2 = mock(JsonNode.class);
-        JsonNode node3 = mock(JsonNode.class);
-        Iterator<JsonNode> iterator = mock(Iterator.class);
-        when(rootNode.iterator()).thenReturn(iterator);
-        // Mock the behavior of the iterator to return JSON nodes when "next" is called
-        when(iterator.hasNext()).thenReturn(true, true, true, false);
-        when(iterator.next()).thenReturn(node1, node2, node3);
-
-        // Mock the behavior of the "path" method to return ticker symbols
-        when(node1.path("symbol")).thenReturn(node1);
-        when(node1.path("symbol").asText()).thenReturn("AAPL");
-
-        when(node2.path("symbol")).thenReturn(node2);
-        when(node2.path("symbol").asText()).thenReturn("GOOG");
-
-        when(node3.path("symbol")).thenReturn(node3);
-        when(node3.path("symbol").asText()).thenReturn("MSFT");
-
-
-        // Mock the behavior of the ObjectMapper to return the JSON structure
-        ObjectMapper objectMapper = mock(ObjectMapper.class);
-        optionsService.setObjectMapper(objectMapper);
-        try {
-            when(objectMapper.readTree(any(File.class))).thenReturn(rootNode);
-
-            List<String> expectedTickers = List.of("AAPL", "GOOG", "MSFT");
-            List<String> actualTickers = optionsService.fetchTickers();
-
-            assertEquals(expectedTickers, actualTickers);
-
-        }catch (IOException e){
-            e.printStackTrace();
-        }
-    }
-
-    @Test
-    public void fetchTickersTest_Fail(){
-        // Mock the behavior of the ObjectMapper to return the JSON structure
-        ObjectMapper objectMapper = mock(ObjectMapper.class);
-        optionsService.setObjectMapper(objectMapper);
-        List<String> actualTickers = new ArrayList<>();
-        try {
-            when(objectMapper.readTree(any(File.class))).thenReturn(null);
-
-            List<String> expectedTickers = List.of("AAPL", "GOOG", "MSFT");
-            actualTickers = optionsService.fetchTickers();
-
-            verify(objectMapper, times(1)).readTree(any(File.class));
-
-        }catch (Exception e){
-
-            assertEquals(0, actualTickers.size());
-        }
     }
 
     @Test
@@ -254,17 +54,15 @@ class OptionsServiceImplTest {
 
         List<OptionsDto> options = optionsService.getOptionsByTicker("AAPL");
         verify(optionsRepository, times(1)).findByTicker("AAPL");
-        assertEquals(2, options.size());
+        Assertions.assertEquals(2, options.size());
     }
 
     @Test
     public void getOptionsByTickerTest_EmptyDB_noCrumb(){
-        optionsService.setCrumb(null);
         when(optionsRepository.findByTicker("AAPL")).thenReturn(Optional.of(List.of()));
 
         List<OptionsDto> options = optionsService.getOptionsByTicker("AAPL");
         verify(optionsRepository, times(1)).findByTicker("AAPL");
-        assertEquals(0, options.size());
+        Assertions.assertEquals(0, options.size());
     }
-
 }
